Convert docker E2E to startup E2E
